Disable the special modulus and export noise (#1)
Does a few pieces to enable performing proofs on ciphertexts:

1. Disables the special modulus with the `disable_special_modulus` parameter to the disable the modulus switching in BFV using the special modulus. This is so that the form of the ciphertexts matches the definition of textbook BFV in a way that it can be shown that the encryption equations hold directly. This is exposed through the longer `Encryptor::encrypt_internal`.
2. The `u` and `e` parameters can be acquired when encrypting a value; this is also required for the later ciphertext correctly constructed proofs.
3. A `StaticPolynomialArray` class is introduced to hold onto polynomial data and easy convert it to and from RNS/BigInt form. The data tied to these arrays is stored on the global memory pool by default. Their size can only be set once; they are meant to store data but not dynamically change size. This can be seen as an array with only one set size (when set), as opposed to the `DynArray` backing other classes like `Ciphertext`.
4. `RNSBase` has been exported through the C interface should it be needed. Currently it is not planned to use this directly in the sunscreen compiler as the `StaticPolynomialArray` class can perform the conversions to and from RNS. However, `RNSBase` has more functionality such as switching modulus bases.

While the outputs seem to be correct, we have not yet performed the full or zero knowledge proof of a correctly constructed ciphertext; more edits may come to disable the special modulus.

Additionally, defining a modulus is now a tad awkward: a modulus array with 5 primes will only use the first four for the public key and for encryption, while the special prime/modulus (the last one in the list) will be ignored. This currently allows the most minimal modifications to SEAL itself. We could potentially investigate deeper and modify more of SEAL to remove the extraneous modulus in the case of the disable special modulus, but this could be quite invasive and may require re-evaluation to determine if the security is equivalent.