Encrypt functions take in seed (#3)
Enables a seed to be optionally set when encrypting a value. This allows for deterministic encryptions for debugging purposes. However, it is disabled in code; to use this feature, you must comment out the random number generation code in `rlwe.cpp`.

Also renames "export_noise" to "export_components", plus some whitespace style changes.